Calculation of coc and point of interest account for camera rotation
Adds support for full camera rotation to the calculate cornea center and
get_point_of_interest methods where a transformation was missing

Changes the final transformation for the point of interest to omit the
last coordinate transform, as both the center_of_cornea_curvature and
the visual_axis_unit_vector are already in WCS, and the z_shift of the
screen plane in WCS is known and already factored into the calculation
of the point of interest.

Changes the calculate_cornea_center_test to update the camera rotation
input parameters to the format used in the integration test, and changes
them to a rotation of 0, as the test truth data assumes.